--- src/main/java/Duke/Command/FindCommand.java
@@ -1,16 +1,0 @@
-package Duke.Command;
-
-import Duke.Exception.DukeException;
-import Duke.TaskList.TaskList;
-
-public class FindCommand extends Command {
-
-    public FindCommand(String searchWord) throws DukeException {
-        super(searchWord);
-    }
-
-    @Override
-    public void callCommand(TaskList taskList) throws DukeException {
-        taskList.findTask(this.input);
-    }
-}
