--- src/main/java/Duke/Exception/DukeException.java
@@ -1,10 +1,0 @@
-package Duke.Exception;
-
-/**
- * Represents an exception that can be thrown by the Duke Application.
- */
-public class DukeException extends java.lang.Exception {
-    public DukeException(String errorMessage) {
-        super(errorMessage);
-    }
-}
