--- src/main/java/Duke/Exception/EmptyByException.java
@@ -1,10 +1,0 @@
-package Duke.Exception;
-
-/**
- * Represents an exception thrown when Deadline task does not have date.
- */
-public class EmptyByException extends DukeException {
-    public EmptyByException(String errorMessage) {
-        super(errorMessage);
-    }
-}
