--- src/main/java/Duke/Exception/EmptyDescriptionException.java
@@ -1,10 +1,0 @@
-package Duke.Exception;
-
-/**
- * Represents an exception thrown when Event, Todo, Deadline task does not have description.
- */
-public class EmptyDescriptionException extends DukeException {
-    public EmptyDescriptionException(String errorMessage) {
-        super(errorMessage);
-    }
-}
