--- src/main/java/Duke/Exception/EmptyEventAtException.java
@@ -1,10 +1,0 @@
-package Duke.Exception;
-
-/**
- * Represents an exception thrown when Event task does not have date.
- */
-public class EmptyEventAtException extends DukeException {
-    public EmptyEventAtException(String errorMessage) {
-        super(errorMessage);
-    }
-}
