--- src/main/java/Duke/Exception/EmptyListException.java
@@ -1,10 +1,0 @@
-package Duke.Exception;
-
-/**
- * Represents an exception thrown when an empty list is requested.
- */
-public class EmptyListException extends DukeException {
-    public EmptyListException(String errorMessage) {
-        super(errorMessage);
-    }
-}
