+++ src/main/java/duke/exception/DukeException.java
@@ -1,0 +1,10 @@
+package duke.exception;
+
+/**
+ * Represents an exception that can be thrown by the Duke Application.
+ */
+public class DukeException extends java.lang.Exception {
+    public DukeException(String errorMessage) {
+        super(errorMessage);
+    }
+}
