+++ src/main/java/duke/exception/EmptyListException.java
@@ -1,0 +1,10 @@
+package duke.exception;
+
+/**
+ * Represents an exception thrown when an empty list is requested.
+ */
+public class EmptyListException extends DukeException {
+    public EmptyListException(String errorMessage) {
+        super(errorMessage);
+    }
+}
