+++ src/main/java/duke/exception/InvalidCommandException.java
@@ -1,0 +1,11 @@
+package duke.exception;
+
+/**
+ * Represents an exception thrown when command given is not recognised.
+ */
+public class InvalidCommandException extends DukeException {
+    public InvalidCommandException(String errorMessage) {
+
+        super(errorMessage);
+    }
+}
