+++ src/main/java/duke/ui/Ui.java
@@ -1,0 +1,18 @@
+package duke.ui;
+
+public class Ui {
+    public Ui() { }
+
+    public void hello() {
+        System.out.println("Hello! Welcome, I'm Duke");
+        System.out.println("What can I do for you?");
+    }
+
+    public void bye() {
+        System.out.print("Bye. Hope to see you again soon!");
+    }
+
+    public void showError(Exception e) {
+        System.out.println(e.getMessage());
+    }
+}
